// scripts/types.d.ts
export type Type_Tuple = {
	type: 'tuple';
	elements: (Element_NamedTupleMember | Type_Intrinsic | Type_Reference)[];
};

export type Type_Literal = {
	type: 'literal';
	value: string | null;
};

export type Type_TemplateLiteral = {
	type: 'templateLiteral';
	head: string | Type_Intrinsic | (Type_Intrinsic | string)[] | (Type_Intrinsic | string)[][];
	tail: string | Type_Intrinsic | (Type_Intrinsic | string)[] | (Type_Intrinsic | string)[][];
};

export type Type_Intrinsic = {
	type: 'intrinsic';
	name: string;
};

export type Type_Array = {
	type: 'array';
	elementType: Type_Intrinsic | Type_Reference;
};

export type Type_Union = {
	type: 'union';
	types: Basic_Types[];
};

export type Type_Query = {
	type: 'query';
	queryType: Basic_Types;
};

export type Type_Predicate = {
	type: 'predicate';
	name: string;
	asserts: boolean;
	targetType: Type_Intrinsic | Type_Reference;
};

export type Type_Intersection = {
	type: 'intersection';
	types: (Type_Reference | Type_Reflection)[];
};

export type Type_Reference = {
	type: 'reference';
	target: number | Target;
	typeArguments?: Basic_Types[];
	name: string;
	package: string;
};

export type Type_TypeOperator = {
	type: 'typeOperator';
	operator: string;
	target?: Basic_Types;
};

export type Type_Reflection = {
	type: 'reflection';
	declaration: Kind_Literal;
};

export type Type_Conditional = {
	type: 'conditional';
	checkType: Basic_Types | Type_IndexedAccess;
	extendsType: Type_TypeOperator;
	trueType: Basic_Types | Type_IndexedAccess;
	falseType: Basic_Types | Type_IndexedAccess;
};

export type Type_IndexedAccess = {
	type: 'indexedAccess';
	indexType: Basic_Types;
	objectType: Basic_Types;
};

export type Element_NamedTupleMember = {
	type: 'namedTupleMember';
	name: string;
	isOptional: boolean;
	element: Basic_Types;
	defaultValue?: string;
};

type Basic_Types =
	| Type_Intrinsic
	| Type_Reference
	| Type_Array
	| Type_Tuple
	| Type_Union
	| Type_Literal
	| Type_TemplateLiteral;

export type All_Types =
	| Basic_Types
	| Type_Intersection
	| Type_Reflection
	| Element_NamedTupleMember
	| Type_Predicate
	| Type_Query
	| Type_Conditional
	| Type_IndexedAccess
	| Type_TypeOperator;
// ...
export type Kind_Literal = BaseProps & {
	kind: ReflectionKind.TypeLiteral;
	children?: Kind_Property[];
	signatures?: Kind_Signature[];
};